+++ docs/setup/release-notes.md
@@ -57,6 +57,7 @@
 #### GeoPandas API
 
 * [<a href='https://github.com/apache/sedona/issues/3269'>GH-3269</a>] - Warn when `geom_equals` or `geom_equals_exact` compares geometry operands with mismatched CRSs
+* [<a href='https://github.com/apache/sedona/issues/3271'>GH-3271</a>] - Preserve per-column CRS state in distributed GeoPandas constructors and select `geometry` or the sole geometry column as active on file reads
 
 ## Sedona 1.9.1